requirement.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-crispy-forms==1.9.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
Faker==4.0.1
idna==2.9
oauthlib==3.1.0
Pillow==7.0.0
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.8
